rip out ListV2 on KeyValueStore
